pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "oresat-c3"
description = "OreSat C3 OLAF app"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "GPL-3.0"}
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Embedded Systems",
]
dependencies = [
    "bitstring",
    "oresat-olaf>=3.0.0",
    "smbus2",
    "spacepackets",
]
dynamic = ["version"]

[project.scripts]
oresat-c3 = "oresat_c3.__main__:main"

[tool.setuptools.dynamic]
version = {attr = "oresat_c3.__version__"}

[tool.setuptools.packages.find]
exclude = ["docs*"] 

[tool.setuptools.package-data]
"*" = ["*.html"]

[tool.black]
line_length = 100

[tool.pylama]
format = "pylint"
skip = "*/.tox/*,*/.env/,*/.git/*,*/.github/*,*/build/*"
linters = "pycodestyle,pyflakes,pylint,mccabe,mypy,radon"
# E402:     Module level import not at top of file
# C901:     Function is too complex
# C0103:    Arguments are not snake_case naming style or too short
# E203:     Whitespace before ':' (black does this)
# R0912:    Too many branches
# R0915:    Too many statements
# R0901:    Function is too complex
# R0914:    Too many local variables
# R1716:    Simplify chained comparison between the operands
# C0413:    Imports not at top of module
# C0206:    Consider iterating dictionaries with .items()
# R1716:    Simplify chained comparison between the operands
# W1514:    Using open without explicitly specifying an encoding
ignore = "E402,C901,C0103,E203,R0912,R0915,R901,R901,R0914,C0413,C0206,R1716,W1514"
max_line_length = 100

[[tool.pylama.files]]
path = "*/__init__.py"
# W0611: imported but unused
ignore = "W0611"

[[tool.mypy.overrides]]
module = "canopen,yaml"
ignore_missing_imports = true

[tool.isort]
profile = "black"
line_length = 100
